Disabled Dynamic Colors for Android R and add member photos for about and data type

+++ CapstoneTestApp/app/src/main/java/com/okta/capstonetestapp/ui/screen/form/TipeForm.kt
@@ -188,7 +188,7 @@ fun TipeForm(
                     Button(
                         onClick = {
                             viewModel.countArea(
-                                inputHarga = inputHarga.toInt(),
+                                inputHarga = inputHarga.toLong(),
                                 selectedYear = selectedYear
                             )
                         },
